site/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
inflect==6.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.43
typing_extensions==4.4.0
Werkzeug==2.2.2
gunicorn==20.1.0
gevent==21.12.0
regex==2022.10.31